Add struct to check launchState
