+++ internal/service/signup.go
@@ -12,7 +12,6 @@
 package service
 
 import (
-	"context"
 	"fmt"
 	"github.com/pkg/errors"
 	entity "github.com/zerotohero-dev/fizz-entity/pkg/data"
@@ -20,8 +19,19 @@ import (
 	"github.com/zerotohero-dev/fizz-env/pkg/env"
 	"github.com/zerotohero-dev/fizz-idm/internal/data"
 	"github.com/zerotohero-dev/fizz-idm/internal/downstream"
-	"github.com/zerotohero-dev/fizz-logging/pkg/log"
 )
+
+
+type Status int
+
+const (
+	Waitlist Status = iota
+	Signup
+)
+
+func (state Status) String() string {
+	return [...]string{"waitlist", "signup"}[state]
+}
 
 func createUnverifiedUser(user entity.User) error {
 	exists, err := data.UserExists(user.Email)
@@ -47,57 +57,7 @@ func createUnverifiedUser(user entity.User) error {
 	return data.CreateUnverifiedUser(user)
 }
 
-func sendEmailVerificationToken(name, email string, emailVerificationToken string) {
-	go func() {
 
-		res, err := downstream.Endpoints().MailerVerification(
-
-			// We are using context.Background() because we do not want this to
-			// cancel prematurely. go-kit cancels the owner context as soon as
-			// the function exits.
-			context.Background(), reqres.RelayEmailVerificationMessageRequest{
-				Email: email,
-				Name:  name,
-				Token: emailVerificationToken,
-			})
-
-		if err != nil {
-			log.Err("Problem sending activation email (%s) (%s)",
-				log.RedactEmail(email), err.Error())
-		}
-
-		er := res.(reqres.RelayEmailVerificationMessageResponse)
-		if er.Err != "" {
-			log.Err("Problem sending activation email (%s) (%s)",
-				log.RedactEmail(email), er.Err)
-		}
-	}()
-}
-
-func sendWaitlistEmail(name, email string) {
-	go func() {
-		res, err := downstream.Endpoints().MailerVerification(
-
-			// We are using context.Background() because we do not want this to
-			// cancel prematurely. go-kit cancels the owner context as soon as
-			// the function exits.
-			context.Background(), reqres.RelayWelcomeMessageRequest{
-				Email: email,
-				Name:  name,
-			})
-
-		if err != nil {
-			log.Err("Problem sending Waitlist email (%s) (%s)",
-				log.RedactEmail(email), err.Error())
-		}
-
-		er := res.(reqres.RelayWelcomeMessageResponse)
-		if er.Err != "" {
-			log.Err("Problem sending Waitlist email (%s) (%s)",
-				log.RedactEmail(email), er.Err)
-		}
-	}()
-}
 
 func (s service) SignUp(user entity.User) error {
 	res, err := downstream.Endpoints().CryptoTokenCreate(
@@ -150,9 +110,9 @@ func (s service) SignUp(user entity.User) error {
 		)
 	}
 
-	launchState := env.FizzEnv().Idm.LaunchState
+	launchState := env.New().Idm.LaunchState
 
-	if launchState == "waitlist" {
+	if launchState == Waitlist.String() {
 		sendWaitlistEmail(user.Name, user.Email)
 	} else {
 		sendEmailVerificationToken(user.Name, user.Email, tr.Token)
